Centered-Neighbor Feature Slick.js
I needed a way to class the slides that were directly adjacent to the center slide to achieve this type of feature: http://i.imgur.com/ymc9FEa.png 

I came up with this rinky dink way of doing it. I'm sure there's a cleaner way to accomplish it, but thought the concept might help others. Good-guy-props to Ken Wheeler for making such a crucial slider and supporting it. 

Thanks!