Remove missing function calls
This might just be my issue, but I'm using libsass to compile and I'm not getting any working paths for the loader or fonts. The only way I could fix them was by removing the "compass" specific functions that were missing. I expected it to fallback but it never did.