Add waitForAnimate option
Default behaviour of slide transitions is to ignore all requests to
advance the slide while animating. This option allows that default
behaviour to be modified.

This option is especially desirable when using the javascript method
slickGoTo(index) because the default behaviour causes calls to
slickGoTo(…) to be dropped if it is in the middle of an animation.