Fix responsive swiping edge cases
There was an issue with swiping the carousel beyond the number
of slides if not using infinite mode.  If you are using responsive
breakpoints, you could end up in a situation where you have fewer
remaining slides than your slidesToScroll value and attempt to go beyond
the slider.  When this happened, the user was brought back to where they
started.

This commit makes it so that if a user attempts to go beyond the end of
the carousel, they will go to the end of the list instead of back to where
they started.  Similarly, if a user attempts to go past the front of the
carousel, they will be taken to the start of the carousel.