Update arrows when refreshing
When moving between responsive breakpoints, the carousel was not always
displaying nav arrows appropriately.  If moving from one breakpoint to
another put you at the end of the carousel, the "next" arrow might
actually look like it's enabled, but if you click it then it suddenly
becomes disabled.

This ensures that the arrows are updated during a
refresh and display properly when switching between breakpoints.