Vertical mode with different slide heights
I was having issues with slides of different heights in vertical mode.
The track would advance the height of the first slide, but the slides
were different heights, which meant they were out of sync with the
track.

This patch adds a quick check in `setDimensions` for vertical mode, and
if set, finds the tallest slide and sets all slides equal to that
height. It also sets the list height accordingly.