Added an alternative `imageLoad` method for the lazyLoad function.
Rather than using the `<img>` tag a `<div>` is used and the image is loaded
into the divs background. The image is then centered and zoomed
to fit the size of the div.

It is important that the dimentions of the div where specified otherwise
nothing will show.

This allows Slick to display a carousel containing images with different
dimentions to display concistenty without having to stretch the image(s)

Example of a slide:

	<div>
		<div class="<your-class-name>" data-lazy="<img-src>"></div>
	</div>