Fix: pauseOnHover re-starts a paused carousel
I needed a way to pause my Slick carousel temporarily. I have `pauseOnHover` set on the carousel. To pause the carousel, I run something like this:

```js
			$(this)
				.slick('slickPause')
				.slick('slickSetOption', 'pauseOnHover', false);

```

However, because the events are registered when the plugin gets initialised, they don't respect changes to the settings. As such, even after running this code I can re-start the carousel by hovering on it then moving my mouse elsewhere.

To resolve this, I moved the if statements inside the event listeners.

This is probably an issue for a handful of other settings, too.

JSFiddle showing the bug: http://jsfiddle.net/15h172h1/