fixes #1311
During initialisation; if a breakpoint which was referenced
in the responsive settings was active , then it would attempt to
refresh the carousel, which would eventually call "asNavFor()"
which would subsequently fail, because the target it references
is not yet a slick instance.

Add a check to see if we're in the initial load, if so there's no reason
to refresh.