Adding in a css style for .slick-arrow.slick-hidden {}
This style is for handling when a breakpoint hits where there's less
slides than we have set to show... arrows should be hidden at this point,
regardless of if they are custom arrows, or dom-elements.

This is non-critical, but non-backwards-compatible change, so should be
added to a minor version with that mentioned.