changing zIndex magic-number in fadeSlilde() and setFade() to a user-declared option variable
PROMPT: our site has a drop down menu whose z-index was being trumped
when we had Slick fade set to true. Rather than adjust and have to
regression test our component, I wanted Slick to be flexible enough to
accommodate pre-existing components’ z-indexes when it’s being
integrated later in a project.

CHANGES: zIndex for currentSlide was being set to “1000”. I added
zIndex to the set of author-able options and set it to 1000 by default.

The extra level of control is non-invasive as the current zIndex
settings are relative to one another. Since none of the other Slick
elements have z-indexing authors couldn’t break Slick itself.

I changed the differences to be single instead of by hundreds since a
user could potentially submit a single-digit zIndex.

Since the setFade() zIndex has to drop by 2, there is a check at the
beginning to make sure the zIndex can never go to 0 or -1 by setting it
to at least 3.

PAIN-POINT: I didn’t put in a check to see if z-index was defined as a
valid number, but I’m not seeing where any other variable is validated
so that seemed okay.