Fixes #1354 - Race condition on Cloned Lazy Load
In some scenarios (one being slidesToShow: 2, slidesToScroll: 1) the
cloned slide would have it's data-lazy removed, but not actually animate
it's opacity to 1, leaving an invisible image inside the cloned slide.

I narrowed it down to the order of code inside the onload() function, and
the code outside of the onload() function. Fixed it by making the loading,
animation and attribute settings synchronous using callbacks from
the .animate() functions.

The setting of opacity: 0; now also has a very fast animation, which makes
the loading of images seem less jarring.