Check for htmlExpr in unload() method
Based on a comment from @nominalaeon I noticed there was a oversight in
the unload method, any arrows which were set in the options as a jquery
selector-string were being removed from the DOM and never replaced in when
a slide was added/removed, or setOption was run, or the slides were
filtered.

Fixed.