Fixes #1248
Because of a condition in `sliderHandler()` checking if `slidesToShow <=
slides` we could never set the slide classes for a nav slider.

Had to split the asNavFor in to two parts, and set the slide classes
in the `slideHandler()` method when `asNavFor` is set and the
`slidesToShow <= slides` -- the other option I didn't want to tackle was
to re-architect the whole `slideHandler` method.

bit hacky. but stable. This architecture  should be reconsidered in 2.0.