Fixed an iOS Bug considering width
On iOS resize Events are triggered when scrolling the page. This means that the slick-slider will get into that part of code where it checks if the old window width (_.windowWidth) has delta with the new one. This also means that the first time scroll happens on iOS the resize will be triggered. 

Furthermore: When you open a scrollable layer and hide all the elements below, the surrounding container of the slick-slider will have no width and height but will try to recaculate based on the fact scrolling happened. This results in unwanted slider results.