Fix tab behavior for accessibility
Tab selection was not possible when tabbing
through an element that was outsideo of the
current view.

This changes the behavior, so that you can
tab through all the slides if they have an
element that can be focused (e.g. <a> link).

When you tab to a slide outside of the current
view, slick will automatically switch its
position so that the slide comes into view.