Add files via upload
I made a few additional changes to improve accessibility of the dots. The aria-label was being read out by screenreaders, but the button text was not. The button text contained helpful contextual information that then only sighted users had knowledge of. By removing the aria-label and setting an aria-describedby attribute for a hidden span containing the "x of Y" supplementary information, that sorted that.

I also removed the role="option" from the button as that's not allowed unless there's a role="listbox" container.