[kenwheeler/slick#3691] Fix extra cloned slides
This for loop iterates over too many slides. In fact it iterates beyond the number of slides but jQuery fails silently once it exceeds the slides. The result is that too many slides get cloned and appended to the end of the track. This previously caused issues fixed with [kenwheeler/slick#3809]. This remaining fix is a minor performance improvement by trimming out the cloned slides that are never visible.